monitoring and debug text improvements; checks;

+++ src/c47/screen.c
@@ -5117,7 +5117,7 @@ static bool_t displayTrueFalse(calcRegister_t regist) {
 
   int16_t refreshScreenCounter = 0;        //JM
 
-  void refreshScreen(uint8_t source) {
+  void refreshScreen(uint16_t source) {
                               #if defined(PC_BUILD) && defined(ANALYSE_REFRESH)
                                 void *callstack[128];
                                 int frames = backtrace(callstack, 128);
@@ -5162,7 +5162,7 @@ static bool_t displayTrueFalse(calcRegister_t regist) {
                                char uuu[100];
                                stringToASCII(indexOfItems[currentMenu() > 0 ? currentMenu() : -currentMenu()].itemSoftmenuName, uuu);
 
-                               printf("   refrsh(%3u): Cnt=%3d %s CM=%2d scr..upd:%3d=%12s=>%26s TI=%4u CL=%s tam:%5i MENUid=%2d:%4i:%s\n",
+                               printf("   refrsh(%5u): Cnt=%3d %s CM=%2d scr..upd:%3d=%12s=>%26s TI=%4u CL=%s tam:%5i MENUid=%2d:%4i:%s\n",
                                   source, refreshScreenCounter++,
                                   (last_CM != calcMode) ? "OVR" : "   ",
                                   calcMode,
