+++ src/c47/screen.h
@@ -155,7 +155,7 @@ char       letteredRegisterName(calcRegister_t regist);
   bool_t   progressHalfSecUpdate_Integer      (uint8_t mode, char *txt, int32_t loop, bool_t clearZ, bool_t clearT, bool_t disp);
   bool_t   monitorExit                        (int32_t *loop, char* str);
   bool_t   checkHalfSec                       (void);
-  void     refreshScreen                      (uint8_t source);
+  void     refreshScreen                      (uint16_t source);
 
   /**
    * Displays a 0 terminated string.
